Super minor bug fix for checkout radio buttons (#88)
is(":checked") no longer works as you expect it to in recent jQuery - using attr("checked") here instead

## How to Test

On the orange theme, if you have 2 or more payment methods, you'll see radio buttons on the left hand side of each one. On the main branch, clicking the labels succeeds in changing payment methods, but slightly annoyingly, clicking the radio button itself will not do anything. 

This update should fix the behavior so that it matches what the code is trying to do.